cloud-composer-trigger-cf/src/requirements.txt
functions-framework==3.2.0
google-auth==2.19.1
requests==2.31.0